BitmaskEnum apply() now returns a Base; cache factory-built Masks
We want to be able to reconstruct defined Mask values from their integral
representation (e.g. during deserialization), so cache the Mask values
that are defined using the Mask factory apply() methods, and bring them
in scope when calling BitmaskEnum.apply() as a fallback to Value id lookup.

Also introduce a convenience function 'transientMask' for creating Mask
instances that we don't necessarily want cached (e.g. the result of some
computation). If an equivalent mask is already in the cache, it will be
used rather than creating a new Mask instance.