Improvements to Myth Protocol API for storage groups (typed return values)
The QUERY_SG_GETFILELIST Myth Protocol command returns a string consisting
of a list of items of the form:

  sgdir::<storage group root>       -or-
    dir::<directory entry>::0       -or-
   file::<file entry>::<file size>::<full path>

Previously we just returned a list of these raw strings from the API call.
However, we would like to parse and return typed results instead.

We create a new marker trait: `StorageGroupInfo', and three subclasses:
`StorageGroupInfoRoot`, `StorageGroupInfoDir` and `StorageGroupInfoFile`
which represent each of the three cases above.

The return type of the querySGGetFileList API is changed from List[String]
to List[StorageGroupInfo]