EventConnection: prevent occasional NPE on eventProtocol usage
Change the eventProtocol member implementation in EventConnection classes
to use lazy vals, to prevent occasional null pointer exceptions due to
initialization order (and multithreading).

It was possible for the eventProtocol member to be used in the
AbstractEventConnection.newEventResponse method before it was
initialized by the subclass override.

This issue was discovered using the RecXfer sample program with debug
logging disabled, as logging slowed down execution sufficiently to
mask the bug.